% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.weights.R
\name{cal.modality.weights}
\alias{cal.modality.weights}
\title{The function to obtain the weights for each modality between each pair of cell types.}
\usage{
cal.modality.weights(
  dat.list,
  cell.type,
  q = 2,
  dnorm,
  subsample.wt = TRUE,
  n.subsample.wt = 200,
  seed = NULL
)
}
\arguments{
\item{dat.list}{a list of data matrices with dimensions reduced.}

\item{cell.type}{a vector of true cell labels.}

\item{q}{the parameter which decides the Lq norm. The default is \code{q = 2} which means using L2 norm.}

\item{dnorm}{a vector that stores the normalization factors, which can be calculated using the \code{cal.dnorm} function}

\item{subsample.wt}{if TRUE, perform subsampling for each cell type. The default is \code{subsample = TRUE}.}

\item{n.subsample.wt}{the number of subsamples for each cell type. The default is \code{n.subsample = 200}. This means cell types with larger than 200
cells will at most use 200 cells to calculate the weights of modalities.}

\item{seed}{the random seed used for subsampling. The default is \code{seed = NULL}.}
}
\value{
a list of weight matrices for each modality.
}
\description{
The function to obtain the weights for each modality between each pair of cell types.
}
