% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIMMO.R
\name{dist.fcn}
\alias{dist.fcn}
\title{The function to obtain the scaling factor and dimension of a data matrix}
\usage{
dist.fcn(dat, sample.idx)
}
\arguments{
\item{dat}{a data matrix whose rows stand for samples and columns stand for PCs.}

\item{sample.idx}{the chosen sample indices which are used for the calculation of scaling factors}
}
\value{
a vector of length 2. The first number is the scaling factor and the second one is the number of PCs.
}
\description{
The function to obtain the scaling factor and dimension of a data matrix
}
