% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.weights.R
\name{get.dat.list}
\alias{get.dat.list}
\title{The function to obtain a list of data matrices from a seurat object}
\usage{
get.dat.list(dat.seu, assay.ls, red.name.ls)
}
\arguments{
\item{dat.seu}{a seurat object with multiple assays. Data preprocessing steps like normalization, high-variable feature selection and
dimension reduction are assumed to be performed.}

\item{assay.ls}{the names of assays in the seurat object. This information can be found using \code{Assays(dat.seu)}.}

\item{red.name.ls}{the names of dimension reductions. This information can be found using \code{names(dat.seu@reductions)}. Note that \code{assay.ls}
and \code{red.name.ls} should match.}
}
\value{
a list of data matrices with dimensions reduced
}
\description{
The function to obtain a list of data matrices from a seurat object
}
