% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIMMO.R
\name{NIMMO}
\alias{NIMMO}
\title{Perform batch integration of multimodal single-cell data using NIMMO.}
\usage{
NIMMO(
  dat.seu,
  assay.ls,
  red.name.ls,
  q = 2,
  subsample = TRUE,
  n.subsample = 1000,
  n.core = 4
)
}
\arguments{
\item{dat.seu}{a seurat object with multiple assays. Data preprocessing steps like normalization, high-variable feature selection and
dimension reduction are assumed to be performed.}

\item{assay.ls}{the names of assays used for integration. This information can be found using \code{Assays(dat.seu)}.}

\item{red.name.ls}{the names of dimension reductions. This information can be found using \code{names(dat.seu@reductions)}. Note that \code{assay.ls}
and \code{red.name.ls} should match.}

\item{q}{the parameter which decides the Lq norm. The default is \code{q = 2} which means using L2 norm to integrate data.}

\item{subsample}{if TRUE, perform subsampling for each data matrix. The default is \code{subsample = TRUE}.}

\item{n.subsample}{the number of subsamples. The default is \code{n.subsample = 1000}. This means each modality will be down-sampled to the
1000 cells if ths sample size is larger than 1000 cells. This subsampling is only used for calculating the scaling factors for each modality.}

\item{n.core}{the number of cores used for the calculation. The default is \code{n.core = 4}}
}
\value{
a seurat object \code{dat} with one new object called \code{nimmo} under \code{dat@reductions}. The new object is a two-dimensional data matrix
obtained by UMAP dimension reduction.
}
\description{
Perform batch integration of multimodal single-cell data using NIMMO.
}
