% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIMMO.R
\name{NIMMO}
\alias{NIMMO}
\title{Perform batch integration of multimodal single-cell data using NIMMO.}
\usage{
NIMMO(dat.seu, assay.ls, red.name.ls, q = 2, sampling.rate = 0.1, n.core = 4)
}
\arguments{
\item{dat.seu}{a seurat object with multiple assays. Data preprocessing steps like normalization, high-variable feature selection and
dimension reduction are assumed to be performed.}

\item{assay.ls}{the names of assays used for integration. This information can be found using \code{Assays(dat.seu)}.}

\item{red.name.ls}{the names of dimension reductions. This information can be found using \code{names(dat.seu@reductions)}. Note that \code{assay.ls}
and \code{red.name.ls} should match.}

\item{q}{the parameter which decides the Lq norm. The default is \code{q = 2} which means using L2 norm to integrate data.}

\item{sampling.rate}{the proportion of cells that are used to calculate scaling factors. The default is \code{sampling.rate = 0.1}.}

\item{n.core}{the number of cores used for the calculation. The default is \code{n.core = 4}}
}
\value{
a seurat object \code{dat} with one new object called \code{nimmo} under \code{dat@reductions}. The new object is a two-dimensional data matrix
obtained by UMAP dimension reduction.
}
\description{
Perform batch integration of multimodal single-cell data using NIMMO.
}
